Refactor and add gaf command
