+++ flatgfa/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod cmds;
 pub mod file;
 pub mod flatgfa;
+pub mod gaf;
 pub mod gfaline;
+pub mod memfile;
 pub mod parse;
 pub mod pool;
 pub mod print;
